+++ neo4j/src/driver/io/bolt/chunk.rs
@@ -101,11 +101,11 @@ impl Deref for Chunk<'_> {
     fn deref(&self) -> &Self::Target {
         match self {
             Chunk::Buffer(buf) => {
-                trace!("C: <RAW> {:02X?}", buf);
+                trace!("C: <RAW> {buf:02X?}");
                 buf
             }
             Chunk::Size(size) => {
-                trace!("C: <RAW> {:02X?}", size);
+                trace!("C: <RAW> {size:02X?}");
                 size
             }
         }
@@ -154,7 +154,7 @@ impl<R: Read> Read for Dechunker<R> {
             if log_enabled!(Level::Trace) {
                 let log_raw = self.chunk_log_raw.as_mut().unwrap();
                 if !log_raw.is_empty() {
-                    trace!("{}]", log_raw);
+                    trace!("{log_raw}]");
                     log_raw.clear();
                 }
                 if self.chunk_size > 0 {
@@ -173,7 +173,7 @@ impl<R: Read> Read for Dechunker<R> {
         if log_enabled!(Level::Trace) && res.is_ok() {
             let log_raw = self.chunk_log_raw.as_mut().unwrap();
             log_raw.push_str(", ");
-            log_raw.push_str(truncate_string(&format!("{:02X?}", buf), 1, 1));
+            log_raw.push_str(truncate_string(&format!("{buf:02X?}"), 1, 1));
         }
         self.chunk_size -= new_buf_size;
         self.error_wrap(res)
@@ -196,7 +196,7 @@ impl<R: Read> Drop for Dechunker<R> {
         if log_enabled!(Level::Trace) {
             let log_raw = self.chunk_log_raw.as_mut().unwrap();
             if !log_raw.is_empty() {
-                trace!("{}]", log_raw);
+                trace!("{log_raw}]");
             }
         }
         if self.chunk_size > 0 && !self.broken {