+++ testkit_backend/src/testkit_backend/session_holder.rs
@@ -142,10 +142,7 @@ impl SessionHolder {
         self.tx_req.send(args.into()).unwrap();
         match self.rx_res.recv().unwrap() {
             CommandResult::RollbackTransaction(result) => result,
-            res => panic!(
-                "expected CommandResult::RollbackTransaction, found {:?}",
-                res
-            ),
+            res => panic!("expected CommandResult::RollbackTransaction, found {res:?}"),
         }
     }
 
@@ -1292,7 +1289,7 @@ impl RecordBuffer {
                     .expect("result stream exhausted above => cannot fail on consume")
                     .map(Arc::new);
             }
-            RecordBuffer::Transaction { .. } => {
+            RecordBuffer::Transaction => {
                 panic!("cannot buffer record stream in transaction")
             }
         }
@@ -1314,7 +1311,7 @@ impl RecordBuffer {
                     Err(e) => records.push_back(Err(Arc::new(e))),
                 }
             }
-            RecordBuffer::Transaction { .. } => {
+            RecordBuffer::Transaction => {
                 panic!("cannot consume record stream in transaction")
             }
         }
@@ -1335,7 +1332,7 @@ impl RecordBuffer {
                         .and_then(|r| r.map(TryInto::try_into).transpose())
                 }
             }
-            RecordBuffer::Transaction { .. } => Err(result_out_of_scope_error()),
+            RecordBuffer::Transaction => Err(result_out_of_scope_error()),
         }
     }
 
@@ -1370,7 +1367,7 @@ impl RecordBuffer {
                     }
                 }
             }
-            RecordBuffer::Transaction { .. } => Err(result_out_of_scope_error()),
+            RecordBuffer::Transaction => Err(result_out_of_scope_error()),
         }
     }