+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ***
 **👏️ Improvements**
  - Reduce network latency in some cases by setting `TCP_NODELAY`.
+ - ⚠️ Reduce the size of `Neo4jError`. The `ServerError` variant is now wrapped in a Box.
 
 
 ## 0.2.0